% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisExpansion_X.R
\name{basisExpansion_X}
\alias{basisExpansion_X}
\title{Basis Expansion for Functional Covariates}
\usage{
basisExpansion_X(
  y = y,
  X = X,
  t.grid = t.grid,
  z = z,
  Bspline_basis = Bspline_basis,
  plot = visualize_basisExpansion_X
)
}
\arguments{
\item{y}{A matrix/array of the response variable.}

\item{X}{A list of functional covariates matrices.}

\item{t.grid}{A numeric vector of the discrete points on which the functions are observed.}

\item{z}{A matrix/array of scalar covariates, default to NULL.}

\item{Bspline_basis}{A basis object created by get_Bspline_basis() function}

\item{plot}{A logical value indicating whether to plot the basis expansion.}
}
\value{
A list containing:
\itemize{
\item X_feature: a list of basis expanded functional covariates
\item y: reordered y
\item z: reordered z
\item ID: the reordered indices
}
}
\description{
This function B-spline performs basis expansion on functional covariates X using the given B-spline basis, with the
option of scaling the Z covariates. The result is a reordered
response and covariates according to the observed domain of each subject.
}
\details{
This function is mainly used for data preprocessing in the fitBASOFR() function, and it is not
intended to be used directly by most users. However, in the case where fitBASOFR() does not yield satisfactory
results, users are advised to utilize basisExpansion_X() to inspect the basis expansion on X.
If the basis expansion fails to capture the trend, consider increasing the number of knots in the num_knots argument in the fitBASOFR() function.
If the basis expansion overfits the X trend, consider decreasing the number of knots in the num_knots argument in the fitBASOFR() function.
}
\examples{
\dontrun{
# simulate data
data = simulate_SOFR(n = 100, p = 2, SNR = 5, num_z_continuous = 1, num_z_dummy = 1,
                       x.grids = seq(0, 1, length.out = 100), beta_types = c('spiky', 'smooth'))
# create B-spline
num_knots=50
knots <- seq(min(data$t.grid), max(data$t.grid),
             (max(data$t.grid)-min(data$t.grid))/(num_knots-1))
Bspline_basis <- get_Bspline_basis(knots)

# use basisExpansion_X to inspect the basis expansion
basisExpansion_X(y=data$y, X=data$X, t.grid=data$t.grid, z_continuous=data$z_continuous,
                 z_dummy= data$z_dummy, Bspline_basis=Bspline_basis, plot= TRUE, scale=TRUE)
}

}
