% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSOFR_helpers.R
\name{FunctionalPredictors}
\alias{FunctionalPredictors}
\title{Simulate functional predictors from Gaussian processes}
\usage{
FunctionalPredictors(
  n,
  l = 0.01,
  x = seq(0, 1, length.out = 201),
  sigma_x = 0.7,
  seasonality = FALSE,
  P = 365/(295 - 1),
  A = 1
)
}
\arguments{
\item{n}{Number of observations}

\item{l}{Characteristic length-scale of the process, default is 0.01}

\item{x}{A vector of values where the functional predictors are observed,
same as the x.grids parameter in the simulate_BASOFR function}

\item{sigma_x}{Standard deviation for the Gaussian process kernel, default is 0.7}

\item{seasonality}{Logical indicating if seasonality should be included in the predictors, default is FALSE}

\item{P}{A parameter controlling the period of the sinusoidal component. Only effective when seasonality == TRUE, default is 365/(295-1)}

\item{A}{Amplitude of the sinusoidal component. Only effective when seasonality == TRUE, default is 1}
}
\value{
A matrix of simulated functional predictors
}
\description{
This function simulates functional predictors with or without seasonality
from Gaussian processes.
}
