% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFuncPred.R
\name{get_func_pred}
\alias{get_func_pred}
\title{Internal function to process functional features}
\usage{
get_func_pred(X.func.feature)
}
\arguments{
\item{X.func.feature}{A list of functional features. Each element of the list is expected
to be a matrix or data frame representing one functional feature.}
}
\value{
A matrix combining all the functional features.
}
\description{
This function processes the functional features by combining them into one matrix.
It's intended for internal use within the package only.
}
\keyword{internal}
