% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_BASOFR.R
\name{Interpret_BASOFR}
\alias{Interpret_BASOFR}
\title{Interpretation of Bayesian Additive Scalar-on-Function Regression Models}
\usage{
Interpret_BASOFR(
  BASOFR_fit,
  y,
  X,
  z = NULL,
  t.grid,
  plot_all_locally_constant = TRUE,
  K = 20,
  eps_level = 0.1
)
}
\arguments{
\item{BASOFR_fit}{A list obtained from the \code{fitBASOFR} function. It should contain the following components:
- \code{beta.post}: Posterior draws of the beta parameter from the Bayesian SOFR model.
- \code{Alpha_continuous_post}: Posterior draws of the Alpha parameter for continuous covariates.
Can be NULL if there are no continuous scalar covariates.
- \code{Alpha_dummy_post}: Posterior draws of the Alpha parameter for dummy covariates.
Can be NULL if there are no dummy scalar covariates.
- \code{intercept_post}: Posterior draws of the intercept from the Bayesian SOFR model.
- \code{sigma.2.post}: Posterior draws of the sigma-squared parameter from the Bayesian SOFR model.}

\item{y}{Vector of observations for the dependent variable.}

\item{X}{Matrix of functional covariate observations.}

\item{z}{Matrix of scalar covariate observations. Can be NULL.}

\item{t.grid}{A numeric vector of the discrete points on which the functions X are observed.}

\item{plot_all_locally_constant}{Logical. If TRUE, all locally constant estimates will be plotted. Defaults to TRUE.}

\item{K}{Number of partitions for the aggregated trajectories. Defaults to 20.}

\item{eps_level}{A numeric value for selecting an acceptable family. It determines the threshold for predictive performance
of models under the Bayesian SOFR model. Defaults to 0.1.}
}
\value{
A list containing:
\itemize{
\item {simplest_acc_family: A list of vectors representing the simplest member in the acceptable family for each functional regression coefficient.}
\item {locally_constant: A list of matrices representing all of the locally constant estimates in the solution path of the fused lasso problem.}
\item {empirical_MSE: A list of vectors containing the empirical MSE of all of the locally constant models in the fused lasso solution path.}
\item {predictive_MSE: A list of matrices representing the predictive MSE of all of the locally constant models in the fused lasso solution path.}
}
}
\description{
This function is designed to simplify the user's interaction when dealing with the BASOFR approach.
It extracts locally constant point estimates from a given Bayesian scalar-on-function regression (SOFR) model.
The output includes plots and a list of results.
}
\examples{
\dontrun{
# simulate the data for BASOFR
data = simulate_BASOFR(n = 1000, p = 3, SNR = 5, num_z_continuous = 0, num_z_dummy = 1,
                       x.grids = seq(0, 1, length.out = 100), beta_types = c('spiky','smooth', 'stepwise'))

# Fit the BASOFR model
BASOFR_fit <- fitBASOFR(y=data$y, X=data$X, t.grid = data$t.grid,
                        z_continuous = data$Z_continuous, z_dummy = data$Z_dummy,
                        S = 1000, burnin = 100)

# Plot the BASOFR fit
plot_BASOFR_fit(BASOFR_fit)

# Check for presence of scalar covariates and create the appropriate `z` matrix
if (!is.null(data$Z_continuous) && !is.null(data$Z_dummy)) {
  z <- cbind(data$Z_continuous, data$Z_dummy)
} else if (is.null(data$Z_continuous) && !is.null(data$Z_dummy)) {
  z <- data$Z_dummy
} else if (!is.null(data$Z_continuous) && is.null(data$Z_dummy)) {
  z <- data$Z_continuous
} else if (is.null(data$Z_continuous) && is.null(data$Z_dummy)) {
  z <- NULL
}

# Use Interpret_BASOFR function
Interpret_BASOFR_result <- Interpret_BASOFR(BASOFR_fit = BASOFR_fit,
                                            y=data$y, X=data$X, z=z,
                                            t.grid=data$t.grid)
}
}
