% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC_BASOFR}
\alias{MCMC_BASOFR}
\title{Run the MCMC for the BASOFR Model}
\usage{
MCMC_BASOFR(
  y,
  X_feature,
  Z,
  Bspline_basis,
  t.grid,
  S = 10000,
  burnin = 10000,
  save = FALSE,
  save_file,
  boundary_reg = TRUE,
  boundary_prior = c(0.01, 0.01),
  z_prior = c(0.01, 0.01),
  sigma2_prior = c(0.01, 0.01)
)
}
\arguments{
\item{y}{The response variable, already processed by the basisExpansion_X() function.}

\item{X_feature}{The functional covariates after basis expansion and multiplication by the J matrix. This list is obtained by the basisExpansion_X() function}

\item{Z}{Scalar covariates, already processed by the basisExpansion_X() function.}

\item{Bspline_basis}{B-spline basis matrix.}

\item{t.grid}{A sequence of discrete points at which the functional covariates are observed.}

\item{S}{The number of iterations in the MCMC chain.}

\item{burnin}{The number of initial MCMC iterations to discard as "burn-in".}

\item{save}{A boolean flag to indicate whether the MCMC output should be saved.}

\item{save_file}{The name of the file where the MCMC output should be saved.}

\item{boundary_reg}{A boolean flag to indicate whether boundary regularization should be implemented.}

\item{boundary_prior}{The prior for the boundary parameter.}

\item{z_prior}{The prior for the z variables.}

\item{sigma2_prior}{The prior for the sigma2 parameter.}
}
\value{
A list containing the MCMC outputs.
}
\description{
This function runs the MCMC for the Bayesian Adaptive Scalar-on-function regression (BASOFR) model.
This function is designed to be used internally by fitBASOFR() function and it is not expected to be called by users directly.
}
\details{
This function fits the Bayesian Adaptive Scalar-on-Function Regression (BASOFR) model via MCMC. The function receives inputs that have already been processed by basisExpansion_X(). The output is a list of MCMC chains for the parameters of interest.
}
\seealso{
\code{\link{fitBASOFR}} for an external function that uses this function internally.
}
