% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{model_selection}
\alias{model_selection}
\title{Model Selection and Visualization for Locally Constant Estimates}
\usage{
model_selection(locallyConstant_fitToAFit, eps_level = 0.1, t.grid, beta.post)
}
\arguments{
\item{locallyConstant_fitToAFit}{An object returned by the \code{get_locallyconstant()} function.}

\item{eps_level}{A numeric value for selecting an acceptable family. It determines the threshold for predictive performance of models under the Bayesian SOFR model.}

\item{t.grid}{A numeric vector of the discrete points on which the functions X are observed.}

\item{beta.post}{A 3D array of the posterior samples of the functional regression coefficients.}
}
\value{
A list of model summaries for each functional regression coefficient. Each summary includes:
- ell_min_in: Best model by in-sample MSE,
- post.dmse: Percent difference in predictive MSE relative to the "best" model,
- dmse: Percent difference in MSE relative to the "best" model,
- ci_dmse: 100(1 - 2*eps_level)\% confidence interval of dmse,
- simplest_accFam_index: Index of the simplest models in the acceptable family,
- simplest_accFam: Matrix of simplest models in the acceptable family.
}
\description{
This function performs the following:
\enumerate{
\item Selects the acceptable family of locally constant estimates for each functional regression coefficient.
\item Plots the simplest member of the acceptable family for each functional regression coefficient.
}
}
\details{
It is a part of Interpret_SOFR function, so typically it's not called directly by the user.
}
\note{
This function is a part of Interpret_SOFR and not typically called by the user directly.
}
\examples{
# This function is usually not called directly, so examples are not provided.
}
