% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSOFR_helpers.R
\name{plot_beta}
\alias{plot_beta}
\title{Plot the true beta functions}
\usage{
plot_beta(beta_list, x.grids)
}
\arguments{
\item{beta_list}{A list of vectors representing the true beta functions.}

\item{x.grids}{A vector of values representing the points at which the
beta functions are evaluated.}
}
\value{
A plot showing all beta functions, each in a different color,
against the corresponding values of x.grids.
}
\description{
Plot the true beta functions
}
\examples{
\dontrun{
sim_data = simulate_SOFR(n=1000, p=3, SNR = 7,
                           num_z_continuous = 0, num_z_dummy = 0,
                           x.grids =seq(0,1,length.out = 201),
                           beta_types = c('spiky', 'smooth', 'spiky'))
plot_beta(beta_list = sim_data$groundTruth$beta, x.grids = sim_data$groundTruth$x.grid)
}
}
