% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_SOFR.R
\name{simulate_SOFR}
\alias{simulate_SOFR}
\title{Simulate data for Scalar-on-Function Regression (SOFR) model}
\usage{
simulate_SOFR(
  n = 1000,
  p = 3,
  SNR = 7,
  num_z_continuous = 0,
  num_z_dummy = 0,
  domain_X = c(0, 1),
  max_obs = 201,
  subject_specific_domain = FALSE,
  beta_types = c("spiky", "smooth", "stepwise")
)
}
\arguments{
\item{n}{The number of observations (default: 1000)}

\item{p}{The number of functional covariates (default: 3)}

\item{SNR}{The signal-to-noise ratio (default: 7)}

\item{num_z_continuous}{The number of continuous control variables (default: 0)}

\item{num_z_dummy}{The number of dummy control variables (default: 0)}

\item{domain_X}{The maximal domain over which the functional covariates X may be observed (default: c(0,1)). The discrete points we get to observe functional X is given by t.grid = seq(domain_X\link{1}, domain_X\link{2}, length.out = max_obs)}

\item{max_obs}{The maximal number of discrete points to observe the functional covariates on domain_X (default: 201)}

\item{subject_specific_domain}{Whether to generate subject-specific domain (default: FALSE)}

\item{beta_types}{The shape of each beta (default: c('spiky', 'smooth', 'stepwise'))}
}
\value{
A list containing the following elements:
\itemize{
\item t.grid: The time grid
\item X: The functional covariates
\item y: The response variable
\item Z_continuous: The continuous control variables
\item Z_dummy: The dummy control variables
\item groundTruth: A list containing the true coefficients, signal-to-noise ratio and SNR for each functional covariate
\item subject_specific_domain: The subject-specific domain vector if \code{subject_specific_domain} is TRUE
}
}
\description{
This function simulates data for a SOFR model, which includes one or more functional covariates,
one or more control variables (both continuous and dummy), and a scalar response. The function
also calculates the signal-to-noise ratio for each functional covariate.
}
\examples{
# Simulate the data for a SOFR model with specific parameters
data = simulate_SOFR(n = 500, p = 2, SNR = 5, num_z_continuous = 1, num_z_dummy = 1,
                     domain_X = c(0,2), max_obs = 150, subject_specific_domain = TRUE,
                     beta_types = c('spiky', 'smooth'))

# Check the output
str(data)

}
